refactor: move and centralize interfaces
