+++ context/AuthContext.tsx
@@ -17,16 +17,7 @@ import { AccessToken, LoginManager } from 'react-native-fbsdk-next';
 
 import api from '@/services/api';
 import AsyncStorage from '@react-native-async-storage/async-storage';
-
-interface AuthContextData {
-  user: AuthResponse['user'] | null;
-  loading: boolean;
-  error: string | null;
-  signIn: (email: string, password: string) => Promise<void>;
-  signUp: (email: string, username: string, password: string) => Promise<void>;
-  signOut: () => Promise<void>;
-  signInWithFacebook: (token: string) => Promise<void>;
-}
+import type { AuthContextData } from '@/interfaces/context/AuthContextData';
 
 export const AuthContext = createContext<AuthContextData>({
   user: null,
