+++ interfaces/components/CategoryGridProps.ts
@@ -1,0 +1,8 @@
+import type { Category, List } from '@/interfaces/services/lists';
+
+export interface CategoryGridProps {
+  categories: Category[];
+  lists: List[];
+  search?: string;
+  onCategoryPress?: (cat: Category) => void;
+}
