+++ interfaces/components/SearchInputProps.ts
@@ -1,0 +1,7 @@
+import type { TextInputProps } from 'react-native';
+
+export interface SearchInputProps extends TextInputProps {
+  value: string;
+  onChangeText: (text: string) => void;
+  style?: any;
+}
