+++ interfaces/forms/RegisterFormData.ts
@@ -1,0 +1,5 @@
+export interface RegisterFormData {
+  email: string;
+  username: string;
+  password: string;
+}
