+++ interfaces/services/categorie.ts
@@ -1,0 +1,6 @@
+import type { Category } from './lists';
+
+export interface TopCategoryOfTheDay {
+  category: Category;
+  hasFilled: boolean;
+}
