+++ interfaces/services/friendship.ts
@@ -1,0 +1,7 @@
+export interface Friendship {
+  id: string;
+  requester: any;
+  addressee: any;
+  status: 'pending' | 'accepted' | 'rejected';
+  createdAt: string;
+}
