+++ interfaces/services/stats.ts
@@ -1,0 +1,16 @@
+export interface CategoryStat {
+  id: string;
+  categoryId?: string;
+  name: string;
+  imageUrl?: string;
+  count: number; // nombre de listes remplies par l'utilisateur
+}
+
+export interface ItemStat {
+  id: string;
+  categoryId: string;
+  item: string;
+  score: number;
+  appearances: number;
+  updatedAt: string;
+}
