+++ services/categorie.ts
@@ -1,10 +1,6 @@
 import api from './api'
-import { Category } from './lists'
-
-export interface TopCategoryOfTheDay {
-  category: Category
-  hasFilled: boolean
-}
+import type { Category } from '@/interfaces/services/lists'
+import type { TopCategoryOfTheDay } from '@/interfaces/services/categorie'
 
 export async function fetchTopCategoryOfTheDay(): Promise<TopCategoryOfTheDay> {
   const { data } = await api.get<TopCategoryOfTheDay>('/categories/top-of-the-day')
